requirements.txt
absl-py==2.1.0
cachetools==5.3.3
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.3.2
CoLT5-attention==0.10.20
cycler==0.11.0
einops==0.6.1
filelock==3.12.2
fonttools==4.38.0
fsspec==2023.1.0
google-auth==2.29.0
google-auth-oauthlib==0.4.6
grpcio==1.62.2
huggingface-hub==0.16.4
idna==3.7
imageio==2.31.2
importlib-metadata==6.7.0
kiwisolver==1.4.5
local-attention==1.9.3
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
mixture-of-experts==0.2.3
networkx==2.6.3
numpy==1.21.6
nvidia-ml-py==12.535.161
nvitop==1.3.2
oauthlib==3.2.2
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
protobuf==3.20.3
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.3.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.3
scikit-image==0.19.3
scipy==1.7.3
seaborn==0.12.2
six==1.16.0
st-moe-pytorch==0.0.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
termcolor==2.3.0
thop==0.1.1.post2209072238
tifffile==2021.11.2
timm==0.9.12
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.66.4
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
